imx6: apalis: move set_blkcnt variable into recovery script
I've added `set_blkcnt` environment variable which is needed for every
`mmc write` command as we need to always specify size in block count.

That `set_blkcnt` variable was previously present in the U-Boot's
default environment, but that needed patching U-Boot every time, so that
patch was dropped with 2020.01 version bump.

Signed-off-by: Petr Štetiar <ynezz@true.cz>